Moved Expense entity
