+++ domain/src/expenses/useCases/updateExpenseUseCase.ts
@@ -1,4 +1,4 @@
-import { buildExpense } from '../entities/factory.js'
+import { buildExpense } from '../../models/factory/expense.js'
 import ExpenseRepository from '../repositories/ExpenseRepository.js'
 
 export default class UpdateExpenseUseCase {
