+++ domain/src/invoices/services/InvoiceService.ts
@@ -1,6 +1,6 @@
 import { v4 as uuid } from 'uuid'
 
-import ExpenseEntity from '../../expenses/entities/ExpensesEntity.js'
+import Expense from '../../models/Expense.js'
 import GuestEntity from '../../guests/entities/GuestEntity.js'
 import { roundMoney } from '../../utils/number.js'
 import InvoiceEntity from '../entities/InvoiceEntity.js'
@@ -25,7 +25,7 @@ class InvoiceService {
     return roundMoney((days / totalDays) * ammount)
   }
 
-  private _calculateExpense (expense: ExpenseEntity, guests: GuestEntity[], invoice: InvoiceEntity): InvoiceEntity {
+  private _calculateExpense (expense: Expense, guests: GuestEntity[], invoice: InvoiceEntity): InvoiceEntity {
     let totalDays = 0
 
     const guestExpenseJoin = guests.map(({ name, period }) => {
