+++ domain/src/models/types.d.ts
